(refactor) AdminMovieService, repository refactoring
