+++ src/main/java/com/grepp/moodlink/app/model/data/movie/MovieService.java
@@ -19,12 +19,8 @@ import java.util.stream.Collectors;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.modelmapper.ModelMapper;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
 @Service
 @RequiredArgsConstructor
@@ -33,8 +29,6 @@ public class MovieService {
 
     private final MovieRepository movieRepository;
     private final GenreRepository genreRepository;
-    private final ModelMapper mapper;
-    private final ImgUploadTemplate imgUploadTemplate;
 
 
     public void saveMovies(List<MovieDto> movieDtos) {
@@ -71,80 +65,6 @@ public class MovieService {
 
             movieRepository.save(movie);
         }
-    }
-
-    public Page<MovieInfoDto> findPaged(Pageable pageable){
-        return movieRepository.findPaged(pageable)
-            .map(MovieInfoDto::toDto);
-    }
-
-    // 모든 장르를 가져옴
-    public List<GenreDto> findAllGenre() {
-        return genreRepository.findAll().stream().map(e-> mapper.map(e, GenreDto.class)).toList();
-    }
-
-    // 영화를 추가함
-    public void addMovie(List<MultipartFile> thumbnail, MovieInfoDto dto) {
-
-        if(movieRepository.existsByTitleAndReleaseDate(dto.getTitle(), dto.getReleaseDate()))
-            throw new CommonException(ResponseCode.DUPLICATED_DATA);
-
-        try {
-
-            uploadImage(thumbnail, dto);
-            Movie movie = mapper.map(dto, Movie.class);
-
-            long count = movieRepository.count();
-            movie.setId("M"+count);
-
-            log.info("{}",movie);
-
-            // 자동으로 안 들어가네
-            movie.setCreatedAt(LocalDate.now());
-
-            movieRepository.save(movie);
-        } catch (IOException e) {
-            throw new CommonException(ResponseCode.INTERNAL_SERVER_ERROR, e);
-        }
-    }
-
-    private void uploadImage(List<MultipartFile> thumbnail, MovieInfoDto dto) throws IOException {
-        if(thumbnail != null){
-            MultipartFile file =  thumbnail.getFirst();
-            String originFileName = file.getOriginalFilename();
-            if (originFileName != null && originFileName.contains(".")) {
-                String ext = originFileName.substring(originFileName.lastIndexOf("."));
-                String renameFileName = UUID.randomUUID().toString() + ext;
-
-                String thumbnailUrl = imgUploadTemplate.uploadImage(file, renameFileName);
-                dto.setThumbnail(thumbnailUrl);
-            }
-        }
-    }
-
-    public MovieInfoDto findById(String id) {
-        return movieRepository.findByIdWithGenre(id).map(MovieInfoDto::toDto).orElse(null);
-    }
-
-    public void updateMovie(List<MultipartFile> thumbnail, MovieInfoDto dto) {
-
-        try {
-            uploadImage(thumbnail, dto);
-
-            // 업데이트
-            movieRepository.updateBook(dto);
-
-            log.info("{}",dto);
-
-
-        } catch (IOException e) {
-            throw new CommonException(ResponseCode.INTERNAL_SERVER_ERROR, e);
-        }
-    }
-
-    @Transactional
-    public void deleteMovie(String id) {
-        movieRepository.findById(id).ifPresent(Movie::unActivated);
     }
 
     @Transactional
